Move resolveAgainst to TypeIdentifier
